+++ test/libsolidity/syntaxTests/types/mapping/array_argument_public.sol
@@ -1,0 +1,6 @@
+contract C {
+    function f(mapping(uint => uint)[] storage) public pure {
+    }
+}
+// ----
+// TypeError: (28-51): Location has to be memory for publicly visible functions (remove the "storage" or "calldata" keyword).