+++ test/libsolidity/syntaxTests/types/mapping/library_array_argument_external.sol
@@ -1,0 +1,6 @@
+library L {
+    function f(mapping(uint => uint)[] storage) external pure {
+    }
+}
+// ----
+// TypeError: (27-50): Type is required to live outside storage.